move the tests of `moss-cache` to `lib.rs`
